Auth
Feature para cadastrar, exibir, e deletar contato por usuário implementada

+++ backend/controllers/contact.controller.js
@@ -12,19 +12,18 @@ const auth = {
 module.exports = {
 
     async register(request, h) {
-
-        if (request.payload === null)
-            return h.response({ message: 'Not JSON' }).code(400)
-
         const userId = request.headers.authorization
 
         try {
             const verify = auth['verify']
             await verify(userId)
         } catch (error) {
-            console.log(error)
             return h.response(error).code(error.code)
         }
+
+        if (request.payload === null)
+            return h.response({ message: 'Not JSON' }).code(400)
+
 
         const contact = new ContactModel({
             name: request.payload.name,
@@ -32,7 +31,6 @@ module.exports = {
             description: request.payload.description,
             userId: userId
         })
-
 
         if (!contact.name)
             return h.response({ message: 'Name is required' }).code(409)
@@ -57,15 +55,39 @@ module.exports = {
 
     },
     async remove(request, h) {
+        const userId = request.headers.authorization
+
         try {
-            await ContactModel.deleteOne({ _id: request.params.contactId })
+            const verify = auth['verify']
+            await verify(userId)
+        } catch (error) {
+            return h.response(error).code(error.code)
+        }
+
+        try {
+            const user = await ContactModel.findOne({ _id: request.params.contactId, userId: userId })
+
+            if (!user)
+                return h.response({}).code(404)
+
+            await ContactModel.deleteOne({ _id: request.params.contactId, userId: userId })
             return h.response({}).code(204)
         } catch (error) {
             return h.response(error).code(500)
         }
     },
     async list(request, h) {
-        const contacts = await ContactModel.find().exec()
+
+        const userId = request.headers.authorization
+
+        try {
+            const verify = auth['verify']
+            await verify(userId)
+        } catch (error) {
+            return h.response(error).code(error.code)
+        }
+
+        const contacts = await ContactModel.find({ userId: userId }).exec()
         return contacts
     }
 }
